refactor(CharacterTile) Extract CharacterTile to it’s own file
